Fragment init file into command chunks
